fixing addressing and bitwaise operations

+++ lib/engine/src/engine.cpp
@@ -1,5 +1,5 @@
 #include "engine.h"
-void EngineControl::load(uint32_t *ar)
+void EngineControl::load(uint16_t *ar)
 {
     for (int i = 0; i < NUMBER_OF_ENGINES; i++)
     {
